Issue 7 solution (#10)
* Experimenting with an SSR aware mode that simply renders through the component.

* Added an SSR mode flag.  When we are running in SSR mode then we simply
render out the wrapped component.  It will receive undefined for both width
and height.

* Added a global flag to control SSR behaviour so that we can replicate the
behaviour on both the client and the server and therefore avoid any React
checksum issues.

* feat(server): Added support for Server Side Rendering.

We now have support (kind of) for SSR, in which case we disable the rendering of the placeholders
and send through null values for width/height to the wrapped components.  Read the docs and tread
carefully.

closes #7